Fixed error-logging issue in PowerShellV2
Issue:
When a cmdlet (run by RunCommand) writes an error to error stream, it's supposed to be logged.
But, when there's a single error, it's not logged in PowerShellV2.

This is because errors captures does not becomes an array.
Errors are captured by $errs, but the logic relies on $errs.Count, which
becomes false in PSv2 case.

This is not the case for PowerShellV3 or later since $errs becomes an array automatically.