Fixed powershell.exe crash when cmdlet timed out
Issue:
When cmdlet timed out, the AsyncWaitHandle was closed before ps instance gets disposed.
When dispoing ps, the cmdlet is also stopped and this seems
to cause powershell to reference the handle and causes powershell.exe to crash.

Fix:
Close the AsyncWaitHandle after cmdlet is stopped.
cmdlet is stopped by disposing ps instance.