Handle error in Stop-Transcript
When the script is stopped in the middle by Ctrl+C, Stop-Transcript in the finally block
returns the following error:

"An error occurred stopping transcription: The host is not currently transcribing."

In this case, the execution stops at the point and the rest of finally block won't get executed.
This happens both in PowerShell & PowerShell ISE.

To avoid this, I've added an error redirect.