Fix for ExchangeOnlineManagement 3.0.0
When ExchangeOnlineManagement 3.0.0 is used to connect to EXO,
no PSSession gets created.

To support scenario (and distinguished from other scenarios), added a
quick fix. This fix is a bit dirty but intended to fix the issue with
minimal changes.

Also include:
- Fix for Save-Object. To save a text file, Out-File is needed instead
of Set-Content. This is a regression when Out-File was replaced by
Set-Content.